Fix collection delete APi, and function to delete, add new function into delete button in chartList.js

+++ frontend/src/pages/ChartListPage.js
@@ -75,26 +75,41 @@ function ChartListPage() {
       if (!userId) {
         throw new Error("Do not have userId");
       }
-      const accessToken= localStorage.getItem("accessToken");
+      // verify
+      let accessToken= localStorage.getItem("accessToken");
+      const refreshToken= localStorage.getItem("refreshToken");
       const decoded = jwtDecode(accessToken);
-      console.log(decoded)
+      const refreshDecoded= jwtDecode(refreshToken);
+      const refreshExpireTime = refreshDecoded.exp*1000;
       const expirationTime = decoded.exp * 1000; // Convert to milliseconds
       const currentTime = Date.now();
       if (currentTime > expirationTime) {
-      localStorage.clear()
-      navigate("/login")
-    }
+        if(currentTime < refreshExpireTime){
+           await axios.post("http://localhost:3000/verify/refresh",{refreshToken:refreshToken, userId: userId})
+               .then(response =>{
+                 console.log(response.data)
+                 localStorage.setItem('accessToken', response.data.newAccessToken)
+               }).catch(error =>{
+                 console.log(error)
+               })
+        } else {
+          localStorage.clear()
+          navigate("/login")
+        }
+      }
       if(!accessToken){
         console.log("access token expried")
         localStorage.clear()
         navigate("/login")
+      }else {
+        accessToken = localStorage.getItem("accessToken")
+        const verify = await axios.post("http://localhost:3000/verify", {access_token: accessToken})
+        if (verify.data.status === "false") {
+          localStorage.clear()
+          navigate("/login")
+        }
       }
-      const verify= await axios.post("http://localhost:3000/verify",{access_token: accessToken})
-      if(verify.data.status==="false"){
-        localStorage.clear()
-        navigate("/login")
-      }
-
+      //verify
 
       const response = await axios.get(
         `http://localhost:3000/chartList/${userId}`
